Fixed poll on gpio sysfs property 'value'
Make this work was actually quite tricky. Indeed, as explained
https://hackage.haskell.org/package/hpio-0.8.0.0/src/src/System/GPIO/Linux/Sysfs/pollSysfs.c
a dummy read is necessary before each poll... yeah, stupid.
PLUS, a lseek to the begining of file after each poll, so that the following dummy read works.
But I now have the plumbery to be notified when a sms arrives, and update the sms list